Split the painter canvas tests into the two lanes
PainterCanvasCanvasTC arrived on master after this series started, and it
drives a real canvas through the Painter dock, so it carries the
NO_LIVE_WINDOW guard and belongs with the other window tests.

Code motion only. _view, _assert_shows and _assert_centered are called
from both sides, so they are duplicated rather than imported across the
directory boundary. The source keeps _StubCanvas and QtGui and loses the
guard, _gui, QtWidgets and os, which no longer have a user there.

Related to #1222
